split split packer class into source and header files; explicitly instantinate packer class for string_buffer type
