+++ runtime/msgpack-serialization.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "msgpack/type.h"
-#include "msgpack/pack.h"
+#include "msgpack/packer.h"
 #include "msgpack/unpacker.h"
 #include "msgpack/unpack_exception.h"
